Even faster transfers
More stats on the UI
FIles saved in directories corresponding to the torrent's name

+++ Torrent.Client/Piece.cs
@@ -32,7 +32,7 @@ namespace Torrent.Client
 
         static public long GetAbsoluteAddress(int index, int offset, int blockSize)
         {
-            return index*blockSize + offset;
+            return (long)index*(long)blockSize + offset;
         }
     }
 }
